+++ src/main/java/net/miarma/api/huertos/services/IncomeService.java
@@ -4,9 +4,10 @@ import java.util.List;
 
 import io.vertx.core.Future;
 import io.vertx.sqlclient.Pool;
-import net.miarma.api.common.QueryParams;
+import net.miarma.api.common.http.QueryParams;
 import net.miarma.api.huertos.dao.IncomeDAO;
 import net.miarma.api.huertos.entities.IncomeEntity;
+import net.miarma.api.util.MessageUtil;
 
 public class IncomeService {
 
@@ -26,10 +27,13 @@ public class IncomeService {
 				.filter(i -> i.getIncome_id().equals(id))
 				.findFirst()
 				.orElse(null);
+			if (income == null) {
+				return Future.failedFuture(MessageUtil.notFound("Income", "in the database"));
+			}
 			return Future.succeededFuture(income);
 		});
 	}
-	
+
 	public Future<List<IncomeEntity>> getUserPayments(Integer memberNumber) {
 		return incomeDAO.getAll().compose(incomes -> {
 			List<IncomeEntity> userPayments = incomes.stream()
@@ -38,11 +42,10 @@ public class IncomeService {
 			return Future.succeededFuture(userPayments);
 		});
 	}
-	
+
 	public Future<Boolean> hasPaid(Integer memberNumber) {
 		return getUserPayments(memberNumber).compose(incomes -> {
-			boolean hasPaid = incomes.stream()
-				.anyMatch(IncomeEntity::isPaid);
+			boolean hasPaid = incomes.stream().anyMatch(IncomeEntity::isPaid);
 			return Future.succeededFuture(hasPaid);
 		});
 	}
@@ -52,10 +55,20 @@ public class IncomeService {
 	}
 
 	public Future<IncomeEntity> update(IncomeEntity income) {
-		return incomeDAO.update(income);
+		return getById(income.getIncome_id()).compose(existing -> {
+			if (existing == null) {
+				return Future.failedFuture(MessageUtil.notFound("Income", "to update"));
+			}
+			return incomeDAO.update(income);
+		});
 	}
 
 	public Future<IncomeEntity> delete(Integer id) {
-		return incomeDAO.delete(id);
+		return getById(id).compose(existing -> {
+			if (existing == null) {
+				return Future.failedFuture(MessageUtil.notFound("Income", "to delete"));
+			}
+			return incomeDAO.delete(id);
+		});
 	}
 }
